requirements.txt
# Project dependencies
langchain-google-genai
langchain-core
langgraph
python-dotenv
pytest
